pve-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.31.0
paho-mqtt==1.6.1
-e ../shared/timeauthority-pkg
